+++ frontend/src/app/modules/campanha/paginas/detalhe-mestre/detalhe-mestre.page.spec.ts
@@ -180,6 +180,7 @@ describe('CampanhaDetalheMestre', () => {
       membroEntrou$: new Subject().asObservable(),
       fichaAlterada$: new Subject().asObservable(),
       fichaVisibilidadeAlterada$: new Subject().asObservable(),
+      fichaCondicoesAlteradas$: new Subject().asObservable(),
       fichaRemovidaDaCampanha$: new Subject().asObservable(),
       rolagemRegistrada$: new Subject().asObservable(),
       rolagemExcluida$: new Subject().asObservable(),